requirements.txt
bax==0.1.9
chex==0.1.0
click==8.0.4
distrax==0.1.1
dm-haiku==0.0.5
einops==0.3.2
jax==0.2.26
jupyterlab==3.3.0
keras==2.7.0
matplotlib==3.5.1
ml-collections==0.1.0
numpy==1.21.5
optax==0.1.0
ray==1.10.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-hub==0.12.0
tensorflow-probability==0.15.0
tqdm==4.62.3
umap-learn==0.5.2